DAOs getAll y save

+++ src/main/java/com/linkedin/learning/otraReunionMas/dao/Dao.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface Dao<T> {
 
-	Optional<T> get(long id);
+	Optional<T> get(int id);
 	List<T> getAll();
 	void save(T t);
 	void update(T t);
